+++ api/internal/core/store/store.go
@@ -42,6 +42,7 @@ type Pagination struct {
 }
 
 type Interface interface {
+	Type() HubKey
 	Get(ctx context.Context, key string) (interface{}, error)
 	List(ctx context.Context, input ListInput) (*ListOutput, error)
 	Create(ctx context.Context, obj interface{}) (interface{}, error)
@@ -64,6 +65,7 @@ type GenericStoreOption struct {
 	KeyFunc    func(obj interface{}) string
 	StockCheck func(obj interface{}, stockObj interface{}) error
 	Validator  Validator
+	HubKey     HubKey
 }
 
 func NewGenericStore(opt GenericStoreOption) (*GenericStore, error) {
@@ -106,7 +108,7 @@ func (s *GenericStore) Init() error {
 		key := ret[i].Key[len(s.opt.BasePath)+1:]
 		objPtr, err := s.StringToObjPtr(ret[i].Value, key)
 		if err != nil {
-			fmt.Fprintln(os.Stderr, "Error occurred while initializing logical store: ", s.opt.BasePath)
+			_, _ = fmt.Fprintln(os.Stderr, "Error occurred while initializing logical store: ", s.opt.BasePath)
 			return err
 		}
 
@@ -142,6 +144,10 @@ func (s *GenericStore) Init() error {
 	return nil
 }
 
+func (s *GenericStore) Type() HubKey {
+	return s.opt.HubKey
+}
+
 func (s *GenericStore) Get(_ context.Context, key string) (interface{}, error) {
 	ret, ok := s.cache.Load(key)
 	if !ok {
@@ -171,8 +177,8 @@ func NewListOutput() *ListOutput {
 }
 
 var defLessFunc = func(i, j interface{}) bool {
-	iBase := i.(entity.BaseInfoGetter).GetBaseInfo()
-	jBase := j.(entity.BaseInfoGetter).GetBaseInfo()
+	iBase := i.(entity.GetBaseInfo).GetBaseInfo()
+	jBase := j.(entity.GetBaseInfo).GetBaseInfo()
 	if iBase.CreateTime != jBase.CreateTime {
 		return iBase.CreateTime < jBase.CreateTime
 	}
@@ -259,7 +265,7 @@ func (s *GenericStore) ingestValidate(obj interface{}) (err error) {
 
 func (s *GenericStore) CreateCheck(obj interface{}) ([]byte, error) {
 
-	if setter, ok := obj.(entity.BaseInfoSetter); ok {
+	if setter, ok := obj.(entity.GetBaseInfo); ok {
 		info := setter.GetBaseInfo()
 		info.Creating()
 	}
@@ -288,7 +294,7 @@ func (s *GenericStore) CreateCheck(obj interface{}) ([]byte, error) {
 }
 
 func (s *GenericStore) Create(ctx context.Context, obj interface{}) (interface{}, error) {
-	if setter, ok := obj.(entity.BaseInfoSetter); ok {
+	if setter, ok := obj.(entity.GetBaseInfo); ok {
 		info := setter.GetBaseInfo()
 		info.Creating()
 	}
@@ -323,8 +329,8 @@ func (s *GenericStore) Update(ctx context.Context, obj interface{}, createIfNotE
 		return nil, fmt.Errorf("key: %s is not found", key)
 	}
 
-	if setter, ok := obj.(entity.BaseInfoGetter); ok {
-		storedGetter := storedObj.(entity.BaseInfoGetter)
+	if setter, ok := obj.(entity.GetBaseInfo); ok {
+		storedGetter := storedObj.(entity.GetBaseInfo)
 		storedInfo := storedGetter.GetBaseInfo()
 		info := setter.GetBaseInfo()
 		info.Updating(storedInfo)
@@ -365,7 +371,7 @@ func (s *GenericStore) StringToObjPtr(str, key string) (interface{}, error) {
 		return nil, fmt.Errorf("json unmarshal failed\n\tRelated Key:\t\t%s\n\tError Description:\t%s", key, err)
 	}
 
-	if setter, ok := ret.(entity.BaseInfoSetter); ok {
+	if setter, ok := ret.(entity.GetBaseInfo); ok {
 		info := setter.GetBaseInfo()
 		info.KeyCompat(key)
 	}