+++ api/internal/core/store/storehub.go
@@ -38,7 +38,8 @@ const (
 	HubKeyGlobalRule   HubKey = "global_rule"
 	HubKeyServerInfo   HubKey = "server_info"
 	HubKeyPluginConfig HubKey = "plugin_config"
-	HubKeyStreamRoute  HubKey = "stream_route"
+  HubKeyProto        HubKey = "proto"
+  HubKeyStreamRoute  HubKey = "stream_route"
 )
 
 var (
@@ -62,6 +63,7 @@ func InitStore(key HubKey, opt GenericStoreOption) error {
 		}
 		opt.Validator = validator
 	}
+	opt.HubKey = key
 	s, err := NewGenericStore(opt)
 	if err != nil {
 		log.Errorf("NewGenericStore error: %s", err)
@@ -202,8 +204,20 @@ func InitStores() error {
 	if err != nil {
 		return err
 	}
-
-	err = InitStore(HubKeyStreamRoute, GenericStoreOption{
+  
+  err = InitStore(HubKeyProto, GenericStoreOption{
+		BasePath: conf.ETCDConfig.Prefix + "/proto",
+		ObjType:  reflect.TypeOf(entity.Proto{}),
+		KeyFunc: func(obj interface{}) string {
+			r := obj.(*entity.Proto)
+      return utils.InterfaceToString(r.ID)
+		},
+	})
+	if err != nil {
+		return err
+	}
+  
+  err = InitStore(HubKeyStreamRoute, GenericStoreOption{
 		BasePath: conf.ETCDConfig.Prefix + "/stream_routes",
 		ObjType:  reflect.TypeOf(entity.StreamRoute{}),
 		KeyFunc: func(obj interface{}) string {
@@ -213,7 +227,7 @@ func InitStores() error {
 	})
 	if err != nil {
 		return err
-	}
-
+  }
+  
 	return nil
 }