+++ web/cypress/integration/route/create-route-with-referer-restriction-form.spec.js
@@ -31,8 +31,14 @@ context('Create and delete route with referer-restriction form', () => {
     notification: '.ant-notification-notice-message',
     notificationCloseIcon: '.ant-notification-close-icon',
     deleteAlert: '.ant-modal-body',
-    whitlist: '#whitelist_0',
+    whitelist: '#whitelist_0',
+    whitelist_1: '#whitelist_1',
+    blacklist: '#blacklist_0',
+    blacklist_1: '#blacklist_1',
     alert: '.ant-form-item-explain-error [role=alert]',
+    newAddWhitelist: '[data-cy=addWhitelist]',
+    newAddBlacklist: '[data-cy=addBlacklist]',
+    passSwitcher: '#bypass_missing',
   };
 
   const data = {
@@ -41,6 +47,9 @@ context('Create and delete route with referer-restriction form', () => {
     weight: 1,
     deleteRouteSuccess: 'Delete Route Successfully',
     submitSuccess: 'Submit Successfully',
+    wrongIp: 'qq@',
+    correctIp: 'apisix-dashboard_1.com',
+    activeClass: 'ant-switch-checked',
   };
 
   beforeEach(() => {
@@ -75,23 +84,33 @@ context('Create and delete route with referer-restriction form', () => {
       .should('be.visible')
       .within(() => {
         cy.get(selector.disabledSwitcher).click();
-        cy.get(selector.checkedSwitcher).should('exist');
+        cy.get(selector.disabledSwitcher).should('have.class', data.activeClass);
+        cy.get(selector.passSwitcher).should('not.have.class', data.activeClass);
       });
 
     // config referer-restriction form without whitelist
-    cy.get(selector.whitlist).click();
-    cy.get(selector.alert).contains('Please Enter whitelist');
+    cy.get(selector.whitelist).click();
     cy.get(selector.drawer).within(() => {
       cy.contains('Submit').click({
         force: true,
       });
     });
     cy.get(selector.notification).should('contain', 'Invalid plugin data');
-    cy.get(selector.notificationCloseIcon).click();
+    cy.get(selector.notificationCloseIcon).click({ multiple: true });
 
     // config referer-restriction form with whitelist
-    cy.get(selector.whitlist).type('127.0.0.1');
+    cy.get(selector.whitelist).type(data.wrongIp);
+    cy.get(selector.whitelist).closest('div').next().children('span').should('exist');
+    cy.get(selector.alert).should('exist');
+    cy.get(selector.whitelist).clear().type(data.correctIp);
     cy.get(selector.alert).should('not.exist');
+
+    cy.get(selector.newAddWhitelist).click();
+    cy.get(selector.whitelist).closest('div').next().children('span').should('exist');
+    cy.get(selector.whitelist_1).closest('div').next().children('span').should('exist');
+    cy.get(selector.whitelist_1).type(data.correctIp);
+    cy.get(selector.alert).should('not.exist');
+
     cy.get(selector.disabledSwitcher).click();
     cy.get(selector.drawer).within(() => {
       cy.contains('Submit').click({
@@ -100,6 +119,41 @@ context('Create and delete route with referer-restriction form', () => {
     });
     cy.get(selector.drawer).should('not.exist');
 
+    // reopen plugin drawer for blacklist test
+    cy.contains('referer-restriction')
+      .parents(selector.pluginCardBordered)
+      .within(() => {
+        cy.get('button').click({
+          force: true,
+        });
+      });
+    cy.get(selector.drawer)
+      .should('be.visible')
+      .within(() => {
+        cy.get(selector.disabledSwitcher).click();
+        cy.get(selector.disabledSwitcher).should('have.class', data.activeClass);
+        cy.get(selector.passSwitcher).should('not.have.class', data.activeClass);
+      });
+    cy.get(selector.blacklist).type(data.correctIp);
+    cy.get(selector.newAddBlacklist).click();
+    cy.get(selector.blacklist_1).type(data.correctIp);
+    cy.get(selector.drawer).within(() => {
+      cy.contains('Submit').click({
+        force: true,
+      });
+    });
+    cy.get(selector.notification).should('contain', 'Invalid plugin data');
+    cy.get(selector.notificationCloseIcon).click({ multiple: true });
+    cy.get(selector.whitelist).closest('div').next().children('span').click();
+    cy.get(selector.whitelist).closest('div').next().children('span').click();
+    cy.get(selector.drawer).within(() => {
+      cy.contains('Submit').click({
+        force: true,
+      });
+    });
+    cy.get(selector.drawer).should('not.exist');
+
+    // create route
     cy.contains('button', 'Next').click();
     cy.contains('button', 'Submit').click();
     cy.contains(data.submitSuccess);