+++ web/cypress/support/commands.js
@@ -33,7 +33,7 @@ Cypress.Commands.add('login', () => {
   });
 });
 
-Cypress.Commands.add('configurePlugins', (cases) => {
+Cypress.Commands.add('configurePlugin', ({ name, cases }) => {
   const timeout = 300;
   const domSelector = {
     name: '[data-cy-plugin-name]',
@@ -46,91 +46,100 @@ Cypress.Commands.add('configurePlugins', (cases) => {
     monacoMode: '[data-cy="monaco-mode"]',
     selectJSON: '.ant-select-dropdown [label=JSON]',
     monacoViewZones: '.view-zones',
+    notification: '.ant-notification-notice-message',
   };
 
+  const shouldValid = cases.shouldValid;
+  const data = cases.data;
+  const type = cases.type;
+
+  if (type === 'consumer') {
+    cy.log('consumer schema case, skipping');
+    return;
+  }
+
   cy.get(domSelector.name, { timeout }).then(function (cards) {
-    [...cards].forEach((card) => {
-      const name = card.innerText;
-      const pluginCases = cases[name] || [];
-      // eslint-disable-next-line consistent-return
-      pluginCases.forEach(({ shouldValid, data, type = '' }) => {
-        if (type === 'consumer') {
-          return true;
-        }
+    let needCheck = false;
+    [...cards].forEach((item) => {
+      if (name === item.innerText) needCheck = true;
+    });
 
-        cy.contains(name)
-          .parents(domSelector.parents)
-          .within(() => {
-            cy.get('button').click({
-              force: true,
-            });
-          });
+    if (!needCheck) {
+      cy.log('non global plugin, skipping');
+      return;
+    }
 
-        // NOTE: wait for the Drawer to appear on the DOM
-        cy.focused(domSelector.drawer).should('exist');
-
-        cy.get(domSelector.monacoMode)
-          .invoke('text')
-          .then((text) => {
-            if (text === 'Form') {
-              cy.wait(5000);
-              cy.get(domSelector.monacoMode).should('be.visible');
-              cy.get(domSelector.monacoMode).click();
-              cy.get(domSelector.selectDropdown).should('be.visible');
-              cy.get(domSelector.selectJSON).click();
-            }
-          });
-
-        cy.get(domSelector.drawer, { timeout }).within(() => {
-          cy.get(domSelector.switch).click({
-            force: true,
-          });
+    cy.contains(name)
+      .parents(domSelector.parents)
+      .within(() => {
+        cy.get('button').click({
+          force: true,
         });
+      });
 
-        cy.get(domSelector.monacoMode)
-          .invoke('text')
-          .then((text) => {
-            if (text === 'Form') {
-              // FIXME: https://github.com/cypress-io/cypress/issues/7306
-              cy.wait(5000);
-              cy.get(domSelector.monacoMode).should('be.visible');
-              cy.get(domSelector.monacoMode).click();
-              cy.get(domSelector.selectDropdown).should('be.visible');
-              cy.get(domSelector.selectJSON).click();
-            }
-          });
-        // edit monaco
-        cy.get(domSelector.monacoViewZones).should('exist').click({ force: true });
-        cy.window().then((window) => {
-          window.monacoEditor.setValue(JSON.stringify(data));
+    // NOTE: wait for the Drawer to appear on the DOM
+    cy.focused(domSelector.drawer).should('exist');
 
-          cy.get(domSelector.drawer, { timeout }).within(() => {
-            cy.contains('Submit').click({
-              force: true,
-            });
-            cy.get(domSelector.drawer).should('not.exist');
-          });
-        });
-
-        if (shouldValid === true) {
-          cy.get(domSelector.drawer).should('not.exist');
-        } else if (shouldValid === false) {
-          cy.get(this.domSelector.notification).should('contain', 'Invalid plugin data');
-
-          cy.get(domSelector.close).should('be.visible').click({
-            force: true,
-            multiple: true,
-          });
-
-          cy.get(domSelector.drawer, { timeout })
-            .invoke('show')
-            .within(() => {
-              cy.contains('Cancel').click({
-                force: true,
-              });
-            });
+    cy.get(domSelector.monacoMode)
+      .invoke('text')
+      .then((text) => {
+        if (text === 'Form') {
+          cy.wait(1000);
+          cy.get(domSelector.monacoMode).should('be.visible');
+          cy.get(domSelector.monacoMode).click();
+          cy.get(domSelector.selectDropdown).should('be.visible');
+          cy.get(domSelector.selectJSON).click();
         }
       });
+
+    cy.get(domSelector.drawer, { timeout }).within(() => {
+      cy.get(domSelector.switch).click({
+        force: true,
+      });
     });
+
+    cy.get(domSelector.monacoMode)
+      .invoke('text')
+      .then((text) => {
+        if (text === 'Form') {
+          // FIXME: https://github.com/cypress-io/cypress/issues/7306
+          cy.wait(1000);
+          cy.get(domSelector.monacoMode).should('be.visible');
+          cy.get(domSelector.monacoMode).click();
+          cy.get(domSelector.selectDropdown).should('be.visible');
+          cy.get(domSelector.selectJSON).click();
+        }
+      });
+    // edit monaco
+    cy.get(domSelector.monacoViewZones).should('exist').click({ force: true });
+    cy.window().then((window) => {
+      window.monacoEditor.setValue(JSON.stringify(data));
+
+      cy.get(domSelector.drawer, { timeout }).within(() => {
+        cy.contains('Submit').click({
+          force: true,
+        });
+        cy.get(domSelector.drawer).should('not.exist');
+      });
+    });
+
+    if (shouldValid === true) {
+      cy.get(domSelector.drawer).should('not.exist');
+    } else if (shouldValid === false) {
+      cy.get(domSelector.notification).should('contain', 'Invalid plugin data');
+
+      cy.get(domSelector.close).should('be.visible').click({
+        force: true,
+        multiple: true,
+      });
+
+      cy.get(domSelector.drawer, { timeout })
+        .invoke('show')
+        .within(() => {
+          cy.contains('Cancel').click({
+            force: true,
+          });
+        });
+    }
   });
 });