+++ web/src/components/Plugin/UI/basic-auth.tsx
@@ -20,6 +20,7 @@ import { Form, Input } from 'antd';
 
 type Props = {
   form: FormInstance;
+  schema: Record<string, any> | undefined;
   ref?: any;
 };
 
@@ -32,13 +33,24 @@ export const FORM_ITEM_LAYOUT = {
   },
 };
 
-const BasicAuth: React.FC<Props> = ({ form }) => {
+const BasicAuth: React.FC<Props> = ({ form, schema }) => {
+  const required: string[] = schema?.required;
   return (
     <Form form={form} {...FORM_ITEM_LAYOUT}>
-      <Form.Item label="username" name="username" required>
+      <Form.Item
+        label="username"
+        name="username"
+        rules={[{ required: required.indexOf('username') > -1 }]}
+        validateTrigger={['onChange', 'onBlur', 'onClick']}
+      >
         <Input></Input>
       </Form.Item>
-      <Form.Item label="password" name="password" required>
+      <Form.Item
+        label="password"
+        name="password"
+        rules={[{ required: required.indexOf('password') > -1 }]}
+        validateTrigger={['onChange', 'onBlur', 'onClick']}
+      >
         <Input></Input>
       </Form.Item>
     </Form>