+++ web/src/pages/Service/locales/zh-CN.ts
@@ -22,6 +22,8 @@ export default {
   'page.service.description':
     '服务由路由中公共的插件配置、上游目标信息组合而成。服务与路由、上游关联，一个服务可对应一组上游节点、可被多条路由绑定。',
   'page.service.fields.name.required': '请输入服务名称',
+  'page.service.fields.hosts': '域名',
+  'page.service.fields.hosts.placeholder': '请输入服务域名',
   'page.service.create': '创建服务',
   'page.service.configure': '配置服务',
 };